requirements.txt
fastapi~=0.103.1
sqlalchemy~=2.0.21
psycopg2-binary
uvicorn[standard]~=0.23.2
starlette~=0.27.0
pydantic~=2.4.0
passlib[bcrypt]~=1.7.4
python-multipart
pyJWT[crypto]~=2.8.0
fastapi-sessions
python-dotenv~=1.0.0